// frontend/bindings/z-server-backup-tools/backend/model/models.ts
export class BackupConfig {
    "host": string;
    "port": number;
    "user": string;
    "password": string;
    "remote_app_dir": string;
    "remote_srv"?: string;
    "remote_state"?: string;
    "remote_source": string;
    "remote_staging"?: string;
    "local_dir": string;
    "max_part_gb": number;
    "part_name_prefix"?: string;
    "task_id"?: string;
    "notify_email"?: string;
    "smtp_host"?: string;
    "smtp_port"?: number;
    "smtp_user"?: string;
    "smtp_password"?: string;
    "known_hosts_file"?: string;

    /** Creates a new BackupConfig instance. */
    constructor($$source: Partial<BackupConfig> = {}) {
        if (!("host" in $$source)) {
            this["host"] = "";
        }
        if (!("port" in $$source)) {
            this["port"] = 0;
        }
        if (!("user" in $$source)) {
            this["user"] = "";
        }
        if (!("password" in $$source)) {
            this["password"] = "";
        }
        if (!("remote_app_dir" in $$source)) {
            this["remote_app_dir"] = "";
        }
        if (!("remote_source" in $$source)) {
            this["remote_source"] = "";
        }
        if (!("local_dir" in $$source)) {
            this["local_dir"] = "";
        }
        if (!("max_part_gb" in $$source)) {
            this["max_part_gb"] = 0;
        }

        Object.assign(this, $$source);
    }

    /**
     * Creates a new BackupConfig instance from a string or object.
     */
    static createFrom($$source: any = {}): BackupConfig {
        let $$parsedSource = typeof $$source === 'string' ? JSON.parse($$source) : $$source;
        return new BackupConfig($$parsedSource as Partial<BackupConfig>);
    }
}